+++ modules/self-hosted/nzbget.nix
@@ -1,4 +1,9 @@
-{ config, lib, pkgs, ... }:
+{
+  config,
+  lib,
+  pkgs,
+  ...
+}:
 
 {
   virtualisation.oci-containers.containers."nzbget" = {
@@ -9,7 +14,7 @@
     };
     user = "3000:3000";
     extraOptions = [
-      "--network=ghostship_net"
+      "--network=container:gluetun"
       "--health-cmd=wget -q --spider --tries=1 --timeout=5 http://127.0.0.1:5001/ || exit 1"
       "--health-interval=30s"
       "--health-timeout=10s"
@@ -30,8 +35,13 @@
   };
 
   systemd.services.podman-nzbget = {
-    after = [ "mnt-share.mount" "init-ghostship-net.service" ];
-    requires = [ "init-ghostship-net.service" ];
+    after = [
+      "mnt-share.mount"
+      "podman-gluetun.service"
+    ];
+    bindsTo = [ "podman-gluetun.service" ];
+    partOf = [ "podman-gluetun.service" ];
+    requires = [ "podman-gluetun.service" ];
     wants = [ "mnt-share.mount" ];
   };